+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- New `rawQuery` API on the `client`, which allows you to performm raw storage queries. It can be useful for reading the `:code` and/or other substrate internal storage values.
+
 ## 1.14.1 - 2025-06-20
 
 ### Fixed